+++ _index.html
@@ -1,50 +1,3 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
-  <meta name="description" content="ウデマエアーカイブは、ガチマッチのウデマエを記録できるサービスだぞ！1戦ごとに勝敗を記録して、グラフを眺めてニヤニヤしたり、苦手なステージを把握したり、連敗の兆候を感じ取ったりしよう！">
-  <meta name="keywords" content="Splatoon,スプラトゥーン,ウデマエ,ガチ,ウデマエアーカイブ,ウデマエア-カイブ">
-  <meta property="og:type" content="website">
-  <meta property="og:site_name" content="ウデマエアーカイブ">
-  <meta property="og:url" content="http://ika.lealog.net">
-  <meta property="og:image" content="http://ika.lealog.net/img/icon.png">
-  <meta property="og:title" content="ウデマエアーカイブ">
-  <meta property="og:description" content="ウデマエアーカイブは、ガチマッチのウデマエを記録できるサービスだぞ！1戦ごとに勝敗を記録して、グラフを眺めてニヤニヤしたり、苦手なステージを把握したり、連敗の兆候を感じ取ったりしよう！">
-  <meta name="twitter:card" content="summary">
-  <meta name="twitter:site" content="@leader22">
-  <meta name="google-site-verification" content="-7YiG_PsYezj78tOs6bd2BOxeaZpIGUnfqq-cQqO-Zo">
-  <title>ウデマエアーカイブ</title>
-  <link rel="apple-touch-icon" sizes="180x180" href="./img/icon.png">
-  <link rel="apple-touch-icon-precomposed" href="./img/icon.png">
-  <link rel="shortcut icon" href="./img/icon.png">
-  <link rel="icon" sizes="192x192" href="./img/icon.png">
-  <link rel="stylesheet" href="./dist/style.css?v=1.20.4">
-</head>
-<body>
-  <div id="jsApp" class="view-intro">
-    <h1 class="ft-ika">ウデマエア-カイブ</h1>
-    <div class="wrap fs-s">
-      <p><span class="ft-ika">ガチマッチ</span>の<span class="ft-ika">ウデマエ</span>を記録できるサービスだぞ！</p>
-      <p>1戦ごとに勝敗を記録して、グラフを眺めてニヤニヤしたり、苦手なステージを把握したり、連敗の兆候を感じ取ったりしよう！</p>
-    </div>
-
-    <h2 class="ft-ika">チュウイ</h2>
-    <div class="wrap fs-s">
-      <p>登録不要・ログインいらずで使えますが、データは直近の{{limit}}戦分のみを、この端末の中に保存します。(localStorage)</p>
-      <p>なので、他の環境にデータは引き継ぎ・引っ越しできません。</p>
-      <p>予告なくサービスの内容を変更したり、終了したりするかもしれません。</p>
-      <p>もしかすると動作しない環境もあるかもしれません・・。</p>
-      <p>まぁそんな適当な感じで良ければ使ってください。</p>
-      <p><span class="ft-ika">イカ、よろしく〜〜〜</span></p>
-    </div>
-
-    <button class="intro-button ft-ika" @click="onClickOk">わかった！</button>
-
-    <footer class="footer">v1.20.4 | &copy; 2015 <a href="https://twitter.com/leader22" target="_blank">leader22</a></footer>
-  </div>
-
-
   <div v-show="!isFirstTime" class="view-app">
     <h1 class="ft-ika">ウデマエア-カイブ</h1>
     <ul class="tab ft-ika"><!--
@@ -315,7 +268,3 @@
 
     <footer class="footer">v1.20.4 | &copy; 2015 <a href="https://twitter.com/leader22" target="_blank">leader22</a></footer>
   </div>
-
-  <script src="./dist/script.js?v=1.20.4"></script>
-</body>
-</html>
