+++ src/script/page/_master.js
@@ -1,0 +1,27 @@
+const React = require('react'); // eslint-disable-line no-unused-vars
+const { Link } = require('react-router');
+
+const Master = ({ Main }) => {
+  return (
+    <div>
+      <h1 className="ft-ika">ウデマエア-カイブ</h1>
+      <ul className="tab ft-ika">
+        <li className="tab-item">
+          <Link to="graph" activeClassName="is-active">キロク</Link>
+        </li>
+        <li className="tab-item">
+          <Link to="user" activeClassName="is-active">セイセキ</Link>
+        </li>
+        <li className="tab-item">
+          <Link to="input" activeClassName="is-active">トウロク</Link>
+        </li>
+        <li className="tab-item">
+          <Link to="others" activeClassName="is-active">ソノタ</Link>
+        </li>
+      </ul>
+      <div>{Main}</div>
+    </div>
+  );
+};
+
+module.exports = Master;
