+++ src/script/page/graph.js
@@ -1,0 +1,9 @@
+const React = require('react'); // eslint-disable-line no-unused-vars
+
+const Graph = () => {
+  return (
+    <div>Graph</div>
+  );
+};
+
+module.exports = Graph;
