+++ src/script/page/user.js
@@ -1,0 +1,9 @@
+const React = require('react'); // eslint-disable-line no-unused-vars
+
+const User = () => {
+  return (
+    <div>User</div>
+  );
+};
+
+module.exports = User;
