+++ src/style/_reset.scss
@@ -39,7 +39,7 @@ p {
 }
 
 a {
-  color: #fff!important;
+  color: #fff;
 }
 
 ul {
